chore: move user model to a models namespace
